pyproject.toml
[project]
name = "voicedrivenfinanceapp"
version = "0.1.0"
description = "Voice-driven smart finance assistant"
readme = "README.md"
requires-python = ">=3.10,<3.12"

dependencies = [
  "alembic==1.13.1",
  "fastapi==0.110.0",
  "ffmpeg-python==0.2.0",
  "gtts==2.5.1",
  "loguru==0.7.2",
  "numpy==1.26.4",
  "openai-whisper==20231117",
  "passlib[bcrypt]==1.7.4",
  "psycopg2-binary==2.9.9",
  "pydantic==2.6.1",
  "python-dateutil==2.8.2",
  "python-jose==3.3.0",
  "python-multipart==0.0.9",
  "python-dotenv==1.2.1",
  "pyttsx3==2.90",
  "redis==5.0.1",
  "regex==2023.12.25",
  "rich==13.7.0",
  "rq==1.15.1",
  "soundfile==0.12.1",
  "sqlalchemy==2.0.25",
  "uvicorn==0.27.1",
  "setuptools>=68.0.0",
  "wheel",
  "typing-extensions>=4.9.0",
  "python-dotenv==1.2.1",
  "langchain>=0.2.17",
  "langchain-openai>=0.1.25",
  "supabase>=2.16.0",
  "transformers>=4.57.3",
  "torch>=2.9.1",
]
